Extract RoomMain from RoomPage
