+++ src/components/Room/index.ts
@@ -1,5 +1,1 @@
 export { default as RoomItem } from "./RoomItem";
-export { default as RoomSettings } from "./RoomSettings/index";
-export { default as RoomQueue } from "./RoomQueue";
-export { default as RoomChat } from "./RoomChat";
-export { default as RoomRules } from "./RoomRules";
